feat(#946): align blog email variants with redesigned blog-subscriber template
Extract buildEmailData + convertContentToHtml into a shared util so the
production batch send and the test-send render the redesigned
blog-subscriber template from identical data.

- send-test-email: now passes custom_message, shop_url, create_url and
  UTM-tagged URLs (previously missing — the two-doors CTAs would render
  with empty/broken links in test emails).
- process-all-batches: refactored to import the shared helpers
  (−123 lines of duplicated logic).
- send-blog-subscribers: polished the admin-ui feed notification copy
  (channel: feed, intentionally plain — not an email).
- plain-text alternative: N/A, emails are HTML-only.
