+++ apps/backend/src/workflows/blogs/send-blog-subscribers/steps/send-test-email.ts
@@ -1,6 +1,9 @@
 import { StepResponse, createStep } from "@medusajs/framework/workflows-sdk"
 import { TestEmailResult } from "../types"
-import { convertTipTapToHtml } from "../utils/tiptap-to-html"
+import {
+  buildEmailData,
+  convertContentToHtml,
+} from "../utils/build-email-data"
 import { sendNotificationEmailWorkflow } from "../../../email/send-notification-email"
 
 export const sendTestEmailStepId = "send-test-email"
@@ -9,12 +12,16 @@ export const sendTestEmailStepId = "send-test-email"
  * This step sends a test email of a blog post to a specified email address.
  * It uses the notification module to send the email and converts TipTap content to HTML.
  *
+ * The test email renders the same redesigned `blog-subscriber` template — and the
+ * same email data payload — as the production subscriber send, so what you see in
+ * the test inbox is exactly what subscribers will receive.
+ *
  * @example
- * const result = sendTestEmailStep({ 
+ * const result = sendTestEmailStep({
  *   email: "test@example.com",
- *   blogData: {...}, 
+ *   blogData: {...},
  *   subject: "Test Blog Email",
- *   customMessage: "This is a test email" 
+ *   customMessage: "This is a test email"
  * })
  */
 interface SendTestEmailInput {
@@ -27,7 +34,6 @@ interface SendTestEmailInput {
 export const sendTestEmailStep = createStep(
   sendTestEmailStepId,
   async (input: SendTestEmailInput, { container }) => {
-    // Validate that email is defined and not empty
     if (!input.email) {
       console.error('Email address is undefined or empty')
       return new StepResponse({
@@ -36,92 +42,40 @@ export const sendTestEmailStep = createStep(
         error: 'Email address is required but was not provided'
       } as TestEmailResult)
     }
-    
+
     console.log(`Sending test email to ${input.email}`)
-    
-    // Use the send-notification-email workflow with blog-subscriber template
-    
+
     try {
-      // Convert TipTap content to HTML
+      // Convert TipTap content to HTML via the shared helper
       let htmlContent = ''
       try {
-        // blogData.content now contains the extracted TipTap content from the blog block
-        const content = input.blogData.content
-        
-        // Handle different content formats
-        if (typeof content === 'object') {
-          // If content is already a parsed object
-          htmlContent = convertTipTapToHtml(content)
-          console.log('Converted TipTap object to HTML')
-        } else if (typeof content === 'string') {
-          // If content is a JSON string
-          if (content.includes('"type":"doc"') || content.startsWith('{')) {
-            try {
-              // Try to parse the JSON string
-              const parsedContent = JSON.parse(content)
-              htmlContent = convertTipTapToHtml(parsedContent)
-              console.log('Converted TipTap JSON string to HTML')
-            } catch (parseError) {
-              // If parsing fails, try to convert the string directly
-              htmlContent = convertTipTapToHtml(content)
-              console.log('Converted TipTap string to HTML (fallback)')
-            }
-          } else {
-            // If it's plain text, use it as is
-            htmlContent = content
-            console.log('Using plain text content')
-          }
-        } else {
-          // Fallback for any other content type
-          htmlContent = String(content || '')
-          console.log('Using fallback string conversion for content')
-        }
+        htmlContent = convertContentToHtml(input.blogData.content)
+        console.log('Converted blog content to HTML')
       } catch (contentError) {
         console.warn(`Failed to convert content to HTML: ${contentError.message}`)
-        // Fall back to a safe default
         htmlContent = String(input.blogData.content || 'No content available')
       }
-      
-      // Prepare email data for the blog-subscriber template
-      const emailData = {
-        // Blog data at root level for template variables
-        blog_title: input.blogData.title,
-        blog_content: htmlContent,
-        blog_url: `${process.env.FRONTEND_URL || ''}${input.blogData.url}`,
-        blog_created_at: input.blogData.created_at,
-        blog_updated_at: input.blogData.updated_at,
-        blog_tags: input.blogData.tags || [],
-        
-        // Person data at root level
-        first_name: "Test",
-        last_name: "User",
-        email: input.email,
-        subscriber_id: "test-user",
-        
-        // Additional template data
-        unsubscribe_url: `${process.env.FRONTEND_URL || ''}/unsubscribe?id=test-user`,
-        website_url: process.env.FRONTEND_URL || '',
-        current_year: new Date().getFullYear().toString(),
-        is_test: true, // Flag to indicate this is a test email
-        
-        // Include nested objects for template compatibility
-        blog: {
-          title: input.blogData.title,
-          content: htmlContent,
-          url: `${process.env.FRONTEND_URL || ''}${input.blogData.url}`,
-          created_at: input.blogData.created_at,
-          updated_at: input.blogData.updated_at,
-          tags: input.blogData.tags || [],
-        },
-        person: {
+
+      // Build the same email data payload used by the production subscriber
+      // send so the test renders the redesigned template identically (UTM-tagged
+      // links, personal note, two-doors CTAs, unsubscribe URL, etc.).
+      const emailData = buildEmailData(
+        {
+          id: "test-user",
+          email: input.email,
           first_name: "Test",
           last_name: "User",
-          email: input.email,
-          id: "test-user",
-        }
-      }
-      
-      // Send email using the new email template workflow
+        },
+        input.blogData,
+        htmlContent,
+        {
+          subject: input.subject,
+          customMessage: input.customMessage,
+        },
+        { isTest: true }
+      )
+
+      // Send email using the email template workflow
       await sendNotificationEmailWorkflow(container).run({
         input: {
           to: input.email,
@@ -129,18 +83,16 @@ export const sendTestEmailStep = createStep(
           data: emailData
         }
       })
-      
+
       console.log(`Successfully sent test email to ${input.email}`)
-      
-      // Return success result
+
       return new StepResponse({
         success: true,
         email: input.email
       } as TestEmailResult)
     } catch (error) {
       console.error(`Failed to send test email to ${input.email}: ${error.message}`)
-      
-      // Return error result
+
       return new StepResponse({
         success: false,
         email: input.email,
